Party Foul: prevent panic when deleting from nil storage provider
- if no storage provider is present, then there's nothing to "Delete" from.  This guards a panic.

+++ queue/queue.go
@@ -220,6 +220,12 @@ func (q *Queue) Schedule(task Task, delay time.Duration) error {
 // Delete removes a task from the queue by its signature
 func (q *Queue) Delete(signature string) error {
 	const location = "queue.Queue.Delete"
+
+	// A memory-only queue has no persistent record to delete
+	if q.storage == nil {
+		return derp.Internal(location, "Must have a storage provider in order to delete tasks")
+	}
+
 	if err := q.storage.DeleteTaskBySignature(signature); err != nil {
 		return derp.Wrap(err, location, "Unable to delete task by signature")
 	}
